+++ cmd/src/batch_exec.go
@@ -142,6 +142,8 @@ func executeBatchSpecInWorkspaces(ctx context.Context, ui *ui.JSONLines, opts ex
 		Timeout:       opts.flags.timeout,
 		KeepLogs:      opts.flags.keepLogs,
 		TempDir:       opts.flags.tempDir,
+		// Temporarily prevent the ability to sending a batch spec with a mount for server-side processing.
+		AllowPathMounts: false,
 	})
 
 	// `src batch exec` uses server-side caching for changeset specs, so we